CRC check completed and OTA logic wip

+++ Bootloader/Core/Inc/crc_helper.h
@@ -8,6 +8,8 @@ extern "C" {
 #include <stdint.h>
 #include <stddef.h>
 
+#include "stm32h7xx_hal.h"
+
 typedef enum {
     CRC_OK = 0,
     CRC_ERROR_INVALID_ADDRESS = -1,
@@ -15,9 +17,8 @@ typedef enum {
     CRC_ERROR_MISMATCH = -3
 } crc_status_t;
 
-uint32_t compute_crc32(CRC_HandleTypeDef* hcrc, uint8_t* data, size_t length);
-crc_status_t verify_crc32(CRC_HandleTypeDef* hcrc, uint8_t* data, size_t length, uint32_t expected_crc);
-
+uint32_t compute_crc32(CRC_HandleTypeDef* hcrc, uint32_t* data, size_t length);
+crc_status_t verify_crc32(CRC_HandleTypeDef* hcrc, uint32_t* data, size_t length, uint32_t expected_crc);
 
 #ifdef __cplusplus
 }
